@@ -63,7 +63,8 @@ func (x *XFile) unzip(zipFile *zip.File) (int64, string, error) {
 	if zipFile.FileInfo().IsDir() {
 		x.Debugf("Writing archived directory: %s", file.Path)
 
-		if err := x.mkDir(file.Path, zipFile.Mode(), zipFile.Modified); err != nil {
+		err := x.mkDir(file.Path, zipFile.Mode(), zipFile.Modified)
+		if err != nil {
 			return 0, file.Path, fmt.Errorf("making zipFile dir: %w", err)
 		}
 
